test: add multi-owner ma tests
