+++ site/pages/reference/account-kit/smart-contracts/functions/getDefaultMultiOwnerLightAccountFactoryAddress.mdx
@@ -1,0 +1,34 @@
+---
+# This file is autogenerated
+title: getDefaultMultiOwnerLightAccountFactoryAddress
+description: Overview of the getDefaultMultiOwnerLightAccountFactoryAddress method
+---
+
+# getDefaultMultiOwnerLightAccountFactoryAddress
+
+Utility method returning the default multi owner light account factory address given a Chain object
+
+## Import
+
+```ts
+import { getDefaultMultiOwnerLightAccountFactoryAddress } from "@account-kit/smart-contracts";
+```
+
+## Parameters
+
+### chain
+
+`Chain`
+
+- a Chain object
+
+### version
+
+`string`
+
+- the version of the light account to get the factory address for
+
+## Returns
+
+`Address`
+an Address for the given chain
